Add high resolution images
Most are from the Project Gutenberg HTML version of the book. "panorama-and-map-of-the-winter-journey.png" and "the-polar-journey.png" were extracted from the NOAA PDF scans. The PDFs unfortunately didn't have any of the plates, which are the only other large images that could benefit from a higher resolution than you get from the PG HTML pages.